refactor(command): move functions into multiple sub-folders
