+++ src/features/discord/messages/functions/commands/checks/discord-strictly-contains-this-command-with-prefix.spec.ts
@@ -1,6 +1,6 @@
 import { createMock } from "ts-auto-mock";
-import { DiscordMessageCommandEnum } from "../../enums/command/discord-message-command.enum";
-import { IDiscordStrictlyContainsThisCommandWithPrefixData } from "../../interfaces/commands/discord-strictly-contains-this-command-with-prefix-data";
+import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
+import { IDiscordStrictlyContainsThisCommandWithPrefixData } from "../../../interfaces/commands/discord-strictly-contains-this-command-with-prefix-data";
 import { discordStrictlyContainsThisCommandWithPrefix } from "./discord-strictly-contains-this-command-with-prefix";
 
 describe(`strictlyContainsThisCommandWithPrefix()`, (): void => {
