+++ src/features/discord/messages/functions/commands/formatters/discord-get-formatted-message.ts
