+++ src/features/discord/messages/functions/commands/getters/discord-get-command-and-first-argument.spec.ts
@@ -1,6 +1,6 @@
 import { createMock } from "ts-auto-mock";
-import { DiscordMessageCommandEnum } from "../../enums/command/discord-message-command.enum";
-import { IDiscordGetCommandAndFirstArgumentData } from "../../interfaces/commands/discord-get-command-and-first-argument-data";
+import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
+import { IDiscordGetCommandAndFirstArgumentData } from "../../../interfaces/commands/discord-get-command-and-first-argument-data";
 import { discordGetCommandAndFirstArgument } from "./discord-get-command-and-first-argument";
 
 describe(`discordGetCommandAndFirstArgument()`, (): void => {
