+++ src/features/discord/messages/functions/commands/getters/discord-get-command-and-prefix.ts
@@ -1,7 +1,7 @@
-import { IDiscordExtractFromCommandCallbackData } from "../../interfaces/commands/discord-extract-from-command-callback-data";
-import { IDiscordGetCommandAndPrefixData } from "../../interfaces/commands/discord-get-command-and-prefix-data";
-import { discordExtractFromCommand } from "./discord-extract-from-command";
-import { discordGetFormattedMessage } from "./discord-get-formatted-message";
+import { IDiscordExtractFromCommandCallbackData } from "../../../interfaces/commands/discord-extract-from-command-callback-data";
+import { IDiscordGetCommandAndPrefixData } from "../../../interfaces/commands/discord-get-command-and-prefix-data";
+import { discordExtractFromCommand } from "../checks/discord-extract-from-command";
+import { discordGetFormattedMessage } from "../formatters/discord-get-formatted-message";
 import { discordGetThisCommandWithPrefix } from "./discord-get-this-command-with-prefix";
 
 function getCommandAndPrefix({
