+++ src/features/discord/messages/functions/commands/getters/discord-get-this-command-with-prefix.spec.ts
@@ -1,6 +1,6 @@
 import { createMock } from "ts-auto-mock";
-import { DiscordMessageCommandEnum } from "../../enums/command/discord-message-command.enum";
-import { IDiscordGetThisCommandWithPrefixData } from "../../interfaces/commands/discord-get-this-command-with-prefix-data";
+import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
+import { IDiscordGetThisCommandWithPrefixData } from "../../../interfaces/commands/discord-get-this-command-with-prefix-data";
 import { discordGetThisCommandWithPrefix } from "./discord-get-this-command-with-prefix";
 
 describe(`discordGetThisCommandWithPrefix()`, (): void => {
