+++ src/features/discord/messages/functions/commands/getters/discord-get-this-command-with-prefix.ts
@@ -1,6 +1,6 @@
 import _ from "lodash";
-import { IDiscordGetThisCommandWithPrefixData } from "../../interfaces/commands/discord-get-this-command-with-prefix-data";
-import { discordGetCommandRegexp } from "./discord-get-command-regexp";
+import { IDiscordGetThisCommandWithPrefixData } from "../../../interfaces/commands/discord-get-this-command-with-prefix-data";
+import { discordGetCommandRegexp } from "../regexp/discord-get-command-regexp";
 import xregexp, { ExecArray } from "xregexp";
 
 export function discordGetThisCommandWithPrefix({
