+++ src/features/discord/messages/functions/commands/regexp/discord-get-command-regexp.spec.ts
@@ -1,6 +1,6 @@
 import { createMock } from "ts-auto-mock";
-import { DiscordMessageCommandEnum } from "../../enums/command/discord-message-command.enum";
-import { IDiscordGetCommandRegexpData } from "../../interfaces/commands/discord-get-command-regexp-data";
+import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
+import { IDiscordGetCommandRegexpData } from "../../../interfaces/commands/discord-get-command-regexp-data";
 import { discordGetCommandRegexp } from "./discord-get-command-regexp";
 import xregexp from "xregexp";
