+++ src/features/discord/messages/services/command/feature/discord-message-command-feature.service.ts
@@ -4,9 +4,9 @@ import { ServiceNameEnum } from "../../../../../../enums/service-name.enum";
 import { ChalkService } from "../../../../../logger/services/chalk/chalk.service";
 import { LoggerService } from "../../../../../logger/services/logger.service";
 import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
-import { discordGetCommandFirstArgument } from "../../../functions/commands/discord-get-command-first-argument";
-import { discordGetCommandFlags } from "../../../functions/commands/discord-get-command-flags";
-import { discordHasThisCommand } from "../../../functions/commands/discord-has-this-command";
+import { discordGetCommandFirstArgument } from "../../../functions/commands/getters/discord-get-command-first-argument";
+import { discordGetCommandFlags } from "../../../functions/commands/getters/discord-get-command-flags";
+import { discordHasThisCommand } from "../../../functions/commands/checks/discord-has-this-command";
 import { IDiscordMessageResponse } from "../../../interfaces/discord-message-response";
 import { IAnyDiscordMessage } from "../../../types/any-discord-message";
 import { DiscordMessageConfigService } from "../../config/discord-message-config.service";
