+++ src/features/discord/messages/services/command/lunch/discord-message-command-lunch.service.ts
@@ -14,7 +14,7 @@ import { DiscordSoniaService } from "../../../../users/services/discord-sonia.se
 import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
 import { DiscordMessageCommandLunchDescriptionEnum } from "../../../enums/command/lunch/discord-message-command-lunch-description.enum";
 import { DiscordMessageCommandLunchTitleEnum } from "../../../enums/command/lunch/discord-message-command-lunch-title.enum";
-import { discordHasThisCommand } from "../../../functions/commands/discord-has-this-command";
+import { discordHasThisCommand } from "../../../functions/commands/checks/discord-has-this-command";
 import { IDiscordMessageResponse } from "../../../interfaces/discord-message-response";
 import { IAnyDiscordMessage } from "../../../types/any-discord-message";
 import { DiscordMessageConfigService } from "../../config/discord-message-config.service";
