+++ src/features/discord/messages/services/command/version/discord-message-command-version.service.ts
@@ -18,7 +18,7 @@ import { DiscordSoniaEmotionalStateEnum } from "../../../../emotional-states/enu
 import { DiscordSoniaEmotionalStateService } from "../../../../emotional-states/services/discord-sonia-emotional-state.service";
 import { DiscordSoniaService } from "../../../../users/services/discord-sonia.service";
 import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
-import { discordHasThisCommand } from "../../../functions/commands/discord-has-this-command";
+import { discordHasThisCommand } from "../../../functions/commands/checks/discord-has-this-command";
 import { IDiscordMessageResponse } from "../../../interfaces/discord-message-response";
 import { IAnyDiscordMessage } from "../../../types/any-discord-message";
 import { DiscordMessageConfigService } from "../../config/discord-message-config.service";
